requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.2
transformers==4.36.2
torch==2.1.2
openai==1.6.1
python-dotenv==1.0.0